Refactor authorization to base recipients class
