+++ lib/mediators/recipients/errors.rb
@@ -1,0 +1,5 @@
+class Mediators
+  module Recipients
+    Forbidden = Class.new(StandardError)
+  end
+end
